eStore/backend/requirements.txt
django==4.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
Pillow==9.5.0